feat(grid): condensed cell kind for masked tool results, thread ordering fix, UI decomposition
- Core: remap compacted thread's ToolResults views to masked replacement
  indexes so into_prompt() resolves to masked content
- History: reverse-map replacement indexes back to originals in GQL output
  so both threads align to the same grid column with different content
- History: sort threads by event-derived creation order (ThreadStarted event
  sequence) instead of relying on HashMap iteration order
- TUI: add CellKind::Condensed (≈) for positions where content differs from
  the owner (masked/simplified tool results)
- TUI: decompose ui.rs into ui/{mod, sidebar, agents, chat_pane, grid}.rs
- TUI: grid navigation skips empty cells, snaps cursor to nearest non-empty
