+++ core/src/agent/session/history.rs
@@ -173,18 +173,32 @@ pub(super) fn build_thread_infos<'a>(
     events: impl Iterator<Item = &'a AgentSessionEvent>,
     current_main_thread: Option<SessionThreadId>,
 ) -> Vec<SessionThreadInfo> {
-    // Build a map of thread_id -> start_reason from session events
-    let start_reasons: std::collections::HashMap<SessionThreadId, ThreadStartReason> = events
-        .filter_map(|e| match e {
-            AgentSessionEvent::ThreadStarted {
-                thread_id,
-                start_reason,
-            } => Some((*thread_id, *start_reason)),
-            _ => None,
-        })
-        .collect();
+    // Build maps of thread_id -> (start_reason, creation_order) from session events.
+    // Event order is the definitive creation order since iter_persisted() comes from
+    // a HashMap whose iteration order is arbitrary.
+    let mut start_reasons: std::collections::HashMap<SessionThreadId, ThreadStartReason> =
+        std::collections::HashMap::new();
+    let mut creation_order: std::collections::HashMap<SessionThreadId, usize> =
+        std::collections::HashMap::new();
+    let mut order_idx = 0usize;
 
-    threads
+    for event in events {
+        if let AgentSessionEvent::ThreadStarted {
+            thread_id,
+            start_reason,
+        } = event
+        {
+            start_reasons.insert(*thread_id, *start_reason);
+            creation_order.insert(*thread_id, order_idx);
+            order_idx += 1;
+        }
+    }
+
+    let mut sorted_threads: Vec<_> = threads.collect();
+    sorted_threads.sort_by_key(|t| creation_order.get(&t.id).copied().unwrap_or(usize::MAX));
+
+    sorted_threads
+        .into_iter()
         .map(|thread| {
             let next_turn = if thread.is_user_turn() {
                 ThreadTurnState::User
@@ -215,6 +229,11 @@ pub(super) fn build_thread_infos<'a>(
 /// Resolves a thread's prompt definition into a list of messages, each carrying
 /// its own block indexes. The global block content list is built by scanning
 /// session events, then each `MessageView` is resolved independently.
+///
+/// For compacted threads whose views contain remapped replacement indexes,
+/// this function reverse-maps them back to the original block indexes in the
+/// GQL output so that both threads reference the same positions — enabling
+/// positional alignment in the thread grid.
 pub(super) fn build_thread_messages(
     prompt_def: PromptDefinition,
     events: &EntityEvents<AgentSessionEvent>,
@@ -222,6 +241,10 @@ pub(super) fn build_thread_messages(
     // Build the global block content list from session events
     // (mirrors the event scan in PromptDefinition::into_prompt).
     let mut all_blocks: Vec<BlockContent> = Vec::new();
+    // Reverse map: replacement block index → original block index.
+    // Used to report original positions in the GQL block_indexes output.
+    let mut reverse_remap: std::collections::HashMap<usize, usize> =
+        std::collections::HashMap::new();
 
     for event in events.iter_all() {
         match event {
@@ -243,6 +266,8 @@ pub(super) fn build_thread_messages(
             }
             AgentSessionEvent::ToolResultsMasked { results, .. } => {
                 for masked in results {
+                    let replacement_idx = all_blocks.len();
+                    reverse_remap.insert(replacement_idx, masked.original_index.index());
                     all_blocks.push(BlockContent::ToolResult(masked.replacement.clone()));
                 }
             }
@@ -254,7 +279,7 @@ pub(super) fn build_thread_messages(
     prompt_def
         .messages
         .iter()
-        .map(|msg_view| resolve_message_view(msg_view, &all_blocks))
+        .map(|msg_view| resolve_message_view(msg_view, &all_blocks, &reverse_remap))
         .collect()
 }
 
@@ -266,7 +291,11 @@ enum BlockContent {
     ToolResult(ToolResultInput),
 }
 
-fn resolve_message_view(view: &MessageView, all_blocks: &[BlockContent]) -> ThreadMessage {
+fn resolve_message_view(
+    view: &MessageView,
+    all_blocks: &[BlockContent],
+    reverse_remap: &std::collections::HashMap<usize, usize>,
+) -> ThreadMessage {
     match view {
         MessageView::User(v) => {
             let block_indexes: Vec<usize> = v.indexes.iter().map(|i| i.index()).collect();
@@ -301,17 +330,28 @@ fn resolve_message_view(view: &MessageView, all_blocks: &[BlockContent]) -> Thre
             }
         }
         MessageView::ToolResults(v) => {
-            let block_indexes: Vec<usize> = v.indexes.iter().map(|i| i.index()).collect();
-            let blocks = block_indexes
+            // Resolve content from the actual view indexes (which may be
+            // replacement indexes for compacted threads).
+            let raw_indexes: Vec<usize> = v.indexes.iter().map(|i| i.index()).collect();
+            let blocks = raw_indexes
                 .iter()
-                .map(|&idx| match &all_blocks[idx] {
-                    BlockContent::ToolResult(result) => ChatHistoryBlock::ToolResult {
+                .map(|&idx| {
+                    let result = match &all_blocks[idx] {
+                        BlockContent::ToolResult(result) => result,
+                        _ => panic!("ToolResults view index does not point to ToolResult"),
+                    };
+                    ChatHistoryBlock::ToolResult {
                         tool_use_id: result.tool_use_id.clone(),
                         content: result.content.clone(),
                         is_error: result.is_error,
-                    },
-                    _ => panic!("ToolResults view index does not point to ToolResult"),
+                    }
                 })
+                .collect();
+            // Reverse-map replacement indexes back to original positions for GQL
+            // output so both threads reference the same column in the grid.
+            let block_indexes = raw_indexes
+                .iter()
+                .map(|&idx| reverse_remap.get(&idx).copied().unwrap_or(idx))
                 .collect();
             ThreadMessage {
                 role: ChatHistoryRole::User,
