+++ drua/src/tui/state.rs
@@ -90,6 +90,8 @@ pub enum CellKind {
     Shared,
     /// Summary block (unique content in a COMPACTION thread).
     Summary(char),
+    /// Condensed — masked/simplified version of the original (e.g. masked tool result).
+    Condensed,
     /// Empty — this thread doesn't reference this position.
     Empty,
 }
@@ -134,6 +136,85 @@ impl ThreadGridState {
 
     pub fn update_visible_cols(&mut self, cols: usize) {
         self.visible_cols = cols;
+    }
+
+    /// Returns true if the cell at (row, col) is non-empty.
+    fn is_non_empty(&self, row: usize, col: usize) -> bool {
+        self.grid
+            .get(row)
+            .and_then(|r| r.get(col))
+            .map(|c| !matches!(c, CellKind::Empty))
+            .unwrap_or(false)
+    }
+
+    /// Snap `cursor_col` to the nearest non-empty cell on the current row.
+    /// Prefers the current column, then searches outward in both directions.
+    /// If the entire row is empty, stays put.
+    pub fn snap_to_nearest_non_empty(&mut self) {
+        let row = self.cursor_row;
+        if row >= self.grid.len() {
+            return;
+        }
+        if self.is_non_empty(row, self.cursor_col) {
+            return;
+        }
+        let cols = self.grid[row].len();
+        // Search outward: check col-1, col+1, col-2, col+2, ...
+        for delta in 1..cols {
+            if self.cursor_col >= delta && self.is_non_empty(row, self.cursor_col - delta) {
+                self.cursor_col -= delta;
+                return;
+            }
+            let right = self.cursor_col + delta;
+            if right < cols && self.is_non_empty(row, right) {
+                self.cursor_col = right;
+                return;
+            }
+        }
+        // Entire row is empty — stay put
+    }
+
+    /// Find the next non-empty column to the right of the current position.
+    pub fn next_non_empty_right(&self) -> Option<usize> {
+        let row = self.cursor_row;
+        if row >= self.grid.len() {
+            return None;
+        }
+        let cols = self.grid[row].len();
+        ((self.cursor_col + 1)..cols).find(|&col| self.is_non_empty(row, col))
+    }
+
+    /// Find the next non-empty column to the left of the current position.
+    pub fn next_non_empty_left(&self) -> Option<usize> {
+        let row = self.cursor_row;
+        if row >= self.grid.len() {
+            return None;
+        }
+        (0..self.cursor_col)
+            .rev()
+            .find(|&col| self.is_non_empty(row, col))
+    }
+
+    /// Find the first non-empty column on the current row.
+    pub fn first_non_empty(&self) -> Option<usize> {
+        let row = self.cursor_row;
+        if row >= self.grid.len() {
+            return None;
+        }
+        self.grid[row]
+            .iter()
+            .position(|c| !matches!(c, CellKind::Empty))
+    }
+
+    /// Find the last non-empty column on the current row.
+    pub fn last_non_empty(&self) -> Option<usize> {
+        let row = self.cursor_row;
+        if row >= self.grid.len() {
+            return None;
+        }
+        self.grid[row]
+            .iter()
+            .rposition(|c| !matches!(c, CellKind::Empty))
     }
 }
 
@@ -394,8 +475,8 @@ impl ScreenState {
 
     pub fn grid_move_right(&mut self) {
         if let Some(ref mut g) = self.thread_view {
-            if !g.positions.is_empty() && g.cursor_col < g.positions.len() - 1 {
-                g.cursor_col += 1;
+            if let Some(col) = g.next_non_empty_right() {
+                g.cursor_col = col;
                 g.ensure_cursor_visible();
             }
         }
@@ -403,8 +484,8 @@ impl ScreenState {
 
     pub fn grid_move_left(&mut self) {
         if let Some(ref mut g) = self.thread_view {
-            if g.cursor_col > 0 {
-                g.cursor_col -= 1;
+            if let Some(col) = g.next_non_empty_left() {
+                g.cursor_col = col;
                 g.ensure_cursor_visible();
             }
         }
@@ -414,6 +495,8 @@ impl ScreenState {
         if let Some(ref mut g) = self.thread_view {
             if !g.threads.is_empty() && g.cursor_row < g.threads.len() - 1 {
                 g.cursor_row += 1;
+                g.snap_to_nearest_non_empty();
+                g.ensure_cursor_visible();
             }
         }
     }
@@ -422,21 +505,25 @@ impl ScreenState {
         if let Some(ref mut g) = self.thread_view {
             if g.cursor_row > 0 {
                 g.cursor_row -= 1;
+                g.snap_to_nearest_non_empty();
+                g.ensure_cursor_visible();
             }
         }
     }
 
     pub fn grid_jump_start(&mut self) {
         if let Some(ref mut g) = self.thread_view {
-            g.cursor_col = 0;
-            g.scroll_col = 0;
+            if let Some(col) = g.first_non_empty() {
+                g.cursor_col = col;
+                g.scroll_col = 0;
+            }
         }
     }
 
     pub fn grid_jump_end(&mut self) {
         if let Some(ref mut g) = self.thread_view {
-            if !g.positions.is_empty() {
-                g.cursor_col = g.positions.len() - 1;
+            if let Some(col) = g.last_non_empty() {
+                g.cursor_col = col;
                 g.ensure_cursor_visible();
             }
         }
